requirements.txt
torch==2.0
torchvision
lightning==1.9.0
tqdm
wandb
plotly
kaleido
torchmetrics
rich
matplotlib
pandas
seaborn
tensorboardx
gdown
hydra-zen
hydra_colorlog
hydra-submitit-launcher
gym==0.26.2
pygame
moviepy
